# tareas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tarea Management API",
        "description": "API for managing tasks, including CRUD operations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List all tasks",
                "description": "Retrieve a list of all tasks in the system",
                "operationId": "4f052e0047edc1db0269dcc2a65ea2a4",
                "responses": {
                    "200": {
                        "description": "A list of tasks"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "description": "Create a new task in the system",
                "operationId": "createTask",
                "requestBody": {
                    "description": "JSON object representing the task to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "example": {
                                "title": "New Task",
                                "description": "This is a test task",
                                "status": "pending"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 1,
                                    "title": "New Task",
                                    "description": "This is a test task",
                                    "status": "pending",
                                    "created_at": "2024-11-20T12:00:00Z",
                                    "updated_at": "2024-11-20T12:00:00Z"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid Input",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Validation error",
                                    "message": "The title field is required."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a specific task",
                "description": "Retrieve details of a task by its ID",
                "operationId": "afb126b8d6a4190d0a216de28fc7f838",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update an existing task",
                "description": "Update details of an existing task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "JSON object representing the task to be updated",
                    "content": {
                        "application/json": {
                            "example": {
                                "title": "Updated Task",
                                "description": "Updated description",
                                "status": "completed"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Task updated successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid Input",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Validation error",
                                    "message": "The title field is required."
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "description": "Remove a task from the system by its ID",
                "operationId": "a5a6287f970b14085a0c28c04b3f2576",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted successfully"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}